recommend_screening4/recommend_screening/requirements.txt
# Core dependencies
pydantic>=2.0.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Text processing & ML
sentence-transformers>=2.2.0
transformers>=4.20.0
torch>=1.12.0

# LangChain
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-core>=0.1.0

# Document processing
pypdf>=3.0.0
python-docx>=0.8.11
Pillow>=9.0.0

# Web framework (for API)
fastapi>=0.100.0
uvicorn>=0.22.0

# Optional database
psycopg2-binary>=2.9.0

# Utilities
python-dotenv>=1.0.0
requests>=2.28.0

# Additional dependencies for CV screening
python-multipart==0.0.6
google-generativeai==0.3.2
langchain-community
PyMuPDF